pendantic

+++ lib/hmlib_files.c
@@ -1,18 +1,13 @@
 #include "hmlib_files.h"
 
-char *hmlib_read(const char *path)
-{
+char *hmlib_read(const char *path) {
     FILE *fp = fopen(path, "r");
-    if (fp == NULL)
-    {
+    if (fp == NULL) {
         printf("file not opened");
         exit(0);
-    }
-    else
-    {
+    } else {
         char ch;
-        while ((ch = fgetc(fp)) != EOF)
-        {
+        while ((ch = fgetc(fp)) != EOF) {
             printf("%c", ch);
         }
     }
@@ -20,16 +15,12 @@ char *hmlib_read(const char *path)
     return NULL;
 }
 
-char *hmlib_buffer_read(const char *path)
-{
+char *hmlib_buffer_read(const char *path) {
     FILE *fp = fopen(path, "r");
-    if (fp == NULL)
-    {
+    if (fp == NULL) {
         printf("file not opened");
         exit(0);
-    }
-    else
-    {
+    } else {
         const int size = 255;
         char buffer[size];
         fgets(buffer, size, fp);
@@ -39,11 +30,9 @@ char *hmlib_buffer_read(const char *path)
     return NULL;
 }
 
-char *hmlib_write(const char *path, const char *content)
-{
+char *hmlib_write(const char *path, const char *content) {
     FILE *fp = fopen(path, "w");
-    if (fp == NULL)
-    {
+    if (fp == NULL) {
         printf("file not opened");
         exit(0);
     }
@@ -52,36 +41,30 @@ char *hmlib_write(const char *path, const char *content)
     return NULL;
 }
 
-size_t hmlib_file_size(const char *path)
-{
+size_t hmlib_file_size(const char *path) {
     FILE *fp = fopen(path, "r");
-    if (fp == NULL)
-    {
+    if (fp == NULL) {
         printf("could not open file");
         exit(1);
     }
     size_t num = 0;
     char ch;
-    while ((ch = fgetc(fp)) != EOF)
-    {
+    while ((ch = fgetc(fp)) != EOF) {
         num++;
     }
     fclose(fp);
     return num;
 }
 
-hmlib_string hmlib_cat(const char *path)
-{
+hmlib_string hmlib_cat(const char *path) {
     size_t size = hmlib_file_size(path);
     FILE *fp = fopen(path, "r");
-    if (fp == NULL)
-    {
+    if (fp == NULL) {
         printf("could not open file");
         exit(1);
     }
     char *content = malloc((size + 1) * sizeof(char));
-    for (int i = 0; i < size; i++)
-    {
+    for (size_t i = 0; i < size; i++) {
         content[i] = fgetc(fp);
     }
     fclose(fp);
